+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 
 - Update exclude types to remove unused options reducing the size of generated code.
+- Add `gitlab-code-quality` error format to print errors in the GitLab Code Quality format
+  for `buf lint` and `buf breaking`.
 
 ## [v1.56.0] - 2025-07-31